src/main/java/com/wn/demo2/properties/BlogProperties.java
package com.wn.demo2.properties;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class BlogProperties {

    @Value(value = "${com.wnspace.blog.name}")
    private String name;

    @Value("${com.wnspace.blog.title}")
    private String title;

    @Value("${com.wnspace.blog.desc}")
    private String desc;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDesc() {
        return desc;
    }

    public void setDesc(String desc) {
        this.desc = desc;
    }
}
